Update to add resource-manager to parser
